ID-325 UI - Application Base Report

+++ source/api/models/ReconciliationYearlySummary.js
@@ -8,7 +8,7 @@ module.exports = {
             type: 'String'
         },
         revenue:{
-            type: 'Integer'
+            type: 'float'
         },
         userId:{
             type:'String'
