+++ source/api/models/RevenueAndTrafficYearlySummary.js
@@ -11,7 +11,7 @@ module.exports = {
             type: 'String'
         },
         revenue:{
-            type: 'Integer'
+            type: 'float'
         },
         viewCount:{
             type:'Integer'
